+++ py2neo/ogm.py
@@ -18,7 +18,7 @@
 """ The ogm module provides Object to Graph Mapping features similar to ORM
 facilities available for relational databases. All functionality is available
 through the :py:class:`Store` class which is bound to a specific
-:py:class:`neo4j.GraphDatabaseService` instance on creation.
+:py:class:`neo4j.Graph` instance on creation.
 
 Conceptually, a mapped object "owns" a single node within the graph along with
 all of that node's outgoing relationships. These features are managed via a
@@ -69,8 +69,8 @@ The code below shows an example of usage::
         def __str__(self):
             return self.name
 
-    graph_db = neo4j.GraphDatabaseService()
-    store = ogm.Store(graph_db)
+    graph = neo4j.Graph()
+    store = ogm.Store(graph)
 
     alice = Person("alice@example.com", "Alice", 34)
     store.save_unique("People", "email", alice.email, alice)
@@ -100,9 +100,9 @@ class NotSaved(ValueError):
 
 class Store(object):
 
-    def __init__(self, graph_db):
-        self.graph_db = graph_db
-        if self.graph_db.supports_optional_match:
+    def __init__(self, graph):
+        self.graph = graph
+        if self.graph.supports_optional_match:
             self.__delete_query = ("START a=node({A}) "
                                    "OPTIONAL MATCH a-[r]-b "
                                    "DELETE r, a")
@@ -223,7 +223,7 @@ class Store(object):
         :param cls: the class of the object to be returned
         :return: a list of `cls` instances
         """
-        index = self.graph_db.get_index(neo4j.Node, index_name)
+        index = self.graph.get_index(neo4j.Node, index_name)
         nodes = index.get(key, value)
         return [self.load(cls, node) for node in nodes]
 
@@ -236,7 +236,7 @@ class Store(object):
         :param cls: the class of the object to be returned
         :return: as instance of `cls` containing the loaded data
         """
-        index = self.graph_db.get_index(neo4j.Node, index_name)
+        index = self.graph.get_index(neo4j.Node, index_name)
         nodes = index.get(key, value)
         if not nodes:
             return None
@@ -283,19 +283,19 @@ class Store(object):
                 props[key] = value
         if hasattr(subj, "__node__"):
             subj.__node__.set_properties(props)
-            query = neo4j.CypherQuery(self.graph_db, "START a=node({A}) "
+            query = neo4j.CypherQuery(self.graph, "START a=node({A}) "
                                                      "MATCH (a)-[r]->(b) "
                                                      "DELETE r")
             query.execute(A=subj.__node__._id)
         else:
-            subj.__node__, = self.graph_db.create(props)
+            subj.__node__, = self.graph.create(props)
         # write rels
         if hasattr(subj, "__rel__"):
-            batch = neo4j.WriteBatch(self.graph_db)
+            batch = neo4j.WriteBatch(self.graph)
             for rel_type, rels in subj.__rel__.items():
                 for rel_props, endpoint in rels:
                     end_node = self._get_node(endpoint)
-                    if not neo4j.familiar(end_node, self.graph_db):
+                    if not neo4j.familiar(end_node, self.graph):
                         raise ValueError(end_node)
                     batch.create((subj.__node__, rel_type, end_node, rel_props))
             batch.run()
@@ -310,7 +310,7 @@ class Store(object):
         :param value: the index value
         :param subj: one or more objects to save
         """
-        index = self.graph_db.get_or_create_index(neo4j.Node, index_name)
+        index = self.graph.get_or_create_index(neo4j.Node, index_name)
         for subj in subj:
             index.add(key, value, self.save(self._get_node(subj)))
 
@@ -323,7 +323,7 @@ class Store(object):
         :param value: the index value
         :param subj: the object to save
         """
-        index = self.graph_db.get_or_create_index(neo4j.Node, index_name)
+        index = self.graph.get_or_create_index(neo4j.Node, index_name)
         node = index.get_or_create(key, value, {})
         self.save(subj, node)
 
@@ -337,5 +337,5 @@ class Store(object):
         self._assert_saved(subj)
         node = subj.__node__
         del subj.__node__
-        neo4j.CypherQuery(self.graph_db,
+        neo4j.CypherQuery(self.graph,
                           self.__delete_query).execute(A=node._id)